feat: 버블차트 API
