+++ src/main/java/gwangjang/server/domain/contents/domain/entity/constant/ApiType.java
@@ -1,0 +1,12 @@
+package gwangjang.server.domain.contents.domain.entity.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ApiType {
+    NAVER,YOUTUBE;
+}
+
+
