+++ src/main/java/gwangjang/server/domain/contents/domain/service/ContentsService.java
@@ -1,0 +1,20 @@
+package gwangjang.server.domain.contents.domain.service;
+import gwangjang.server.domain.contents.application.dto.res.ContentsRes;
+import gwangjang.server.domain.contents.domain.entity.constant.ApiType;
+import gwangjang.server.global.annotation.DomainService;
+import reactor.core.publisher.Mono;
+
+import java.util.List;
+
+@DomainService
+public interface ContentsService {
+    public Mono<Void> saveYoutubeContent(String[] search);
+    public List<ContentsRes> getContents(ApiType type);
+    List<ContentsRes> getContentsTitle(String issue);
+    List<ContentsRes> getKeywordAndType(String Keyword, ApiType apiType);
+    ContentsRes getContentsById(Integer contentsId);
+    List<ContentsRes> getContentLikeCount();
+    List<ContentsRes> findContentsByLoginId(String loginId);
+
+}
+
