+++ src/main/java/gwangjang/server/domain/contents/presentation/constant/ContentsResponseMessage.java
@@ -1,0 +1,14 @@
+package gwangjang.server.domain.contents.presentation.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ContentsResponseMessage {
+
+    GET_CONTENTS_SUCCESS("컨텐츠 조회 완료");
+
+    private final String message;
+
+}
