+++ src/main/java/gwangjang/server/domain/contents/util/KomoranUtils.java
@@ -1,0 +1,4 @@
+package gwangjang.server.domain.contents.util;
+
+public class KomoranUtils {
+}
