+++ src/main/java/gwangjang/server/domain/like/application/dto/res/ContentLikeRes.java
@@ -1,8 +1,5 @@
 package gwangjang.server.domain.like.application.dto.res;
 
-import gwangjang.server.domain.morpheme.domain.entity.Contents;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.*;
 
 @Getter
