+++ src/main/java/gwangjang/server/domain/like/domain/service/LikeService.java
@@ -3,8 +3,8 @@ package gwangjang.server.domain.like.domain.service;
 import gwangjang.server.domain.like.application.dto.res.ContentLikeRes;
 import gwangjang.server.domain.like.domain.entity.ContentLike;
 import gwangjang.server.domain.like.domain.repository.LikeRepository;
-import gwangjang.server.domain.morpheme.domain.entity.Contents;
-import gwangjang.server.domain.morpheme.domain.repository.ContentsRepository;
+import gwangjang.server.domain.contents.domain.entity.Contents;
+import gwangjang.server.domain.contents.domain.repository.ContentsRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
