--- src/main/java/gwangjang/server/domain/morpheme/domain/entity/Issue.java
@@ -1,21 +1,0 @@
-package gwangjang.server.domain.morpheme.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class Issue {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "issue_id")
-    private Long id;
-    private String issueTitle;
-    @ManyToOne
-    @JoinColumn(name = "topic_id")
-    private Topic topic;
-}
