--- src/main/java/gwangjang/server/domain/morpheme/domain/entity/Topic.java
@@ -1,21 +1,0 @@
-package gwangjang.server.domain.morpheme.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class Topic {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "topic_id")
-    private Long id;
-
-    private String topicTitle;
-
-
-}
