--- src/main/java/gwangjang/server/domain/morpheme/domain/entity/constant/ApiType.java
@@ -1,12 +1,0 @@
-package gwangjang.server.domain.morpheme.domain.entity.constant;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ApiType {
-    NAVER,YOUTUBE;
-}
-
-
