--- src/main/java/gwangjang/server/domain/morpheme/domain/repository/ContentsCustomRepository.java
@@ -1,14 +1,0 @@
-package gwangjang.server.domain.morpheme.domain.repository;
-
-import gwangjang.server.domain.morpheme.application.dto.res.ContentsDataRes;
-import gwangjang.server.domain.morpheme.domain.entity.Contents;
-
-import java.util.List;
-
-public interface ContentsCustomRepository {
-    List<ContentsDataRes> getContentsByIssueId(String issue);
-    List<Contents> findAllOrderByLikeCountDesc();
-    List<Contents> findContentsByLoginId(String loginId);
-    void updateContentsImageUrl(Integer contentsId, String newImageUrl);
-
-}
