--- src/main/java/gwangjang/server/domain/morpheme/domain/repository/ContentsRepository.java
@@ -1,16 +1,0 @@
-package gwangjang.server.domain.morpheme.domain.repository;
-
-import gwangjang.server.domain.morpheme.domain.entity.Contents;
-import gwangjang.server.domain.morpheme.domain.entity.constant.ApiType;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ContentsRepository extends JpaRepository<Contents, Integer>,ContentsCustomRepository{
-
-    List<Contents> findByType(ApiType type);
-    List<Contents> findByIssueTitleLike(String issue);
-    List<Contents> findByKeywordLikeAndTypeOrderByPubDateDesc(String issueTitle, ApiType type);
-}
