--- src/main/java/gwangjang/server/domain/morpheme/presentation/constant/ContentsResponseMessage.java
@@ -1,14 +1,0 @@
-package gwangjang.server.domain.morpheme.presentation.constant;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ContentsResponseMessage {
-
-    GET_CONTENTS_SUCCESS("컨텐츠 조회 완료");
-
-    private final String message;
-
-}
