--- src/main/java/gwangjang/server/domain/morpheme/util/KomoranUtils.java
@@ -1,4 +1,0 @@
-package gwangjang.server.domain.morpheme.util;
-
-public class KomoranUtils {
-}
